Rework context inputs additions
